Move phoneNumber tests into different folder to seperate them from Adnan's tests
